ms_mint_conc renamed to ms_conc
