+++ setup.py
@@ -9,15 +9,15 @@ install_requires = [
 
 config = {
     'description': 'compute concentration by using standard curves',
-    'author': 'luis-ponce and soren-wacker',
-    'url': 'https://github.com/luis-ponce',
-    'download_url': 'https://github.com/luis-ponce/ms_mint_conc',
+    'author': 'luis-ponce and sorenwacker',
+    'url': 'https://github.com/LewisResearchGroup/ms-conc',
+    'download_url': 'https://github.com/LewisResearchGroup/ms-conc',
     'author_email': 'luisfponcinho@gmail.com',
     'version': '0.0.1',
     'install_requires': install_requires,
     'packages': find_packages(),
     'scripts': [],
-    'name': 'ms_mint_conc'
+    'name': 'ms_conc'
 }
 
 setup(**config)
