+++ streamlit_app.py
@@ -1,8 +1,8 @@
-from ms_mint_conc import calibration_curves as cc
-from ms_mint_conc import ConcentrationEstimator as CE
-from ms_mint_conc import AppState as AS
-from ms_mint_conc import SessionState
-from ms_mint_conc.SessionState import get
+from ms_conc import calibration_curves as cc
+from ms_conc import ConcentrationEstimator as CE
+from ms_conc import AppState as AS
+from ms_conc import SessionState
+from ms_conc.SessionState import get
 import matplotlib.pyplot as plt
 import os
 import pandas as pd
